moving all svg to src folder to fix svg imports, fixing the inject-styles import issue
