+++ rollup.config.js
@@ -18,21 +18,41 @@ export default [
       }
     ],
     external: [
+      /@babel\/runtime/,
+      /style-inject/,
       ...Object.keys(pkg.peerDependencies || {}),
       ...Object.keys(pkg.devDependencies || {}),
-      /@babel\/runtime/
     ],
     plugins: [
       nodeResolve({
         extensions: ['.js',  '.jsx', '.css']
       }),
+      postcss({
+        modules: true,
+        plugins: [autoprefixer()],
+        minimize: true
+      }),
+      {
+        name: 'Fix style-inject import plugin',
+        generateBundle: (options, bundle) => {
+          Object.entries(bundle).forEach(entry => {
+            // early return if the file we're currently looking at doesn't need to be acted upon by this plugin
+            if (!entry[0].match(/.*(.css.js)$/)) {
+              return;
+            }
+
+            // this line only runs for .scss.js files, which were generated by the postcss plugin.
+            // depending on the use-case, the relative path to style-inject might need to change
+            bundle[entry[0]].code = entry[1].code.replace(
+              '../node_modules/style-inject/dist/style-inject.es.js',
+              'style-inject',
+            );
+          });
+        },
+      },
       babel({
         babelHelpers: 'runtime',
         exclude: 'node_modules/**',
-      }),
-      postcss({
-        plugins: [autoprefixer()],
-        minimize: true
       }),
       terser(), // minifies generated bundles
       rename({
@@ -44,8 +64,8 @@ export default [
         map: (name) => name.replace('src/components', 'components'),
       }),
       rename({
-        include: 'public/**/*.jsx',
-        map: (name) => name.replace('public/svg', 'svg'),
+        include: 'src/svg/**/*.jsx',
+        map: (name) => name.replace('src/svg', 'svg'),
       })
     ]
   }
