+++ src/svg/ArrowCloseIcon.jsx
