+++ src/svg/ArrowOpenIcon.jsx
