+++ src/svg/CloseIcon.jsx
