+++ src/svg/SearchIcon.jsx
