+++ src/svg/SpotifyLogo.jsx
