+++ src/svg/SpotifyLogoText.jsx
